2023.05.17

+++ src/main/java/com/matsinger/barofishserver/data/CurationController.java
@@ -1,5 +1,8 @@
 package com.matsinger.barofishserver.data;
 
+import com.matsinger.barofishserver.jwt.JwtService;
+import com.matsinger.barofishserver.jwt.TokenAuthType;
+import com.matsinger.barofishserver.jwt.TokenInfo;
 import com.matsinger.barofishserver.product.Product;
 import com.matsinger.barofishserver.product.ProductService;
 import com.matsinger.barofishserver.utils.Common;
@@ -12,11 +15,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 @RestController
 @RequiredArgsConstructor
@@ -31,72 +30,74 @@ public class CurationController {
     private final Common util;
 
     private final S3Uploader s3;
+    private final JwtService jwt;
 
     // GET
     @Description("큐레이션 목록")
     @GetMapping("/")
-    public ResponseEntity<CustomResponse> selectCurationList() {
-        CustomResponse res = new CustomResponse();
+    public ResponseEntity<CustomResponse<List<Curation>>> selectCurationList() {
+        CustomResponse<List<Curation>> res = new CustomResponse();
         try {
             List<Curation> curations = curationService.selectCurations();
             res.setData(Optional.ofNullable(curations));
             return ResponseEntity.ok(res);
-        } catch (Error error) {
-            res.setIsSuccess(false);
-            res.setErrorMsg(error.getMessage());
-            return ResponseEntity.ok(res);
+        } catch (Exception e) {
+            return res.defaultError(e);
         }
     }
 
     @Description("큐레이션 상품 목록")
     @GetMapping("/{id}/products")
-    public ResponseEntity<CustomResponse> selectCurationProducts(@PathVariable("id") Long id) {
-        CustomResponse res = new CustomResponse();
+    public ResponseEntity<CustomResponse<List<Product>>> selectCurationProducts(@PathVariable("id") Long id) {
+        CustomResponse<List<Product>> res = new CustomResponse();
         try {
             List<Product> products = curationService.selectCurationProducts(id);
             res.setData(Optional.ofNullable(products));
             return ResponseEntity.ok(res);
-        } catch (Error error) {
-            res.setIsSuccess(false);
-            res.setErrorMsg(error.getMessage());
-            return ResponseEntity.ok(res);
+        } catch (Exception e) {
+            return res.defaultError(e);
         }
     }
 
     //POST
     @Description("큐레이션 추가")
     @PostMapping(value = "/add", consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.MULTIPART_FORM_DATA_VALUE})
-    public ResponseEntity<CustomResponse> createCuration(@RequestPart(value = "data") Curation curation,
-                                                         @RequestPart(value = "image") MultipartFile file) {
-        CustomResponse res = new CustomResponse();
+    public ResponseEntity<CustomResponse<Curation>> createCuration(@RequestHeader(value = "Authorization") Optional<String> auth,
+                                                                   @RequestPart(value = "image") MultipartFile file,
+                                                                   @RequestPart(value = "shortName") String shortName,
+                                                                   @RequestPart(value = "title") String title,
+                                                                   @RequestPart(value = "description") String description,
+                                                                   @RequestPart(value = "type") CurationType type) {
+        CustomResponse<Curation> res = new CustomResponse();
+        Optional<TokenInfo> tokenInfo = jwt.validateAndGetTokenInfo(Set.of(TokenAuthType.ADMIN), auth);
+        if (tokenInfo == null) return res.throwError("인증이 필요합니다.", "FORBIDDEN");
         try {
-            String shortName = util.validateString(curation.getShortName(), 20L, "약어");
+            Curation curation = new Curation();
+            shortName = util.validateString(shortName, 20L, "약어");
             curation.setShortName(shortName);
-            String title = util.validateString(curation.getTitle(), 100L, "제목");
+            title = util.validateString(title, 100L, "제목");
             curation.setTitle(title);
-            String description = util.validateString(curation.getDescription(), 200L, "설명");
+            description = util.validateString(description, 200L, "설명");
             curation.setDescription(description);
             String image = s3.upload(file, new ArrayList<>(Arrays.asList("curation")));
             curation.setImage(image);
+            curation.setType(type);
             Curation data = curationService.add(curation);
             res.setData(Optional.ofNullable(data));
             return ResponseEntity.ok(res);
-        } catch (Error error) {
-            res.setIsSuccess(false);
-            res.setErrorMsg(error.getMessage());
-            return ResponseEntity.ok(res);
-        } catch (IOException e) {
-            res.setIsSuccess(false);
-            res.setErrorMsg(e.getMessage());
-            return ResponseEntity.ok(res);
+        } catch (Exception e) {
+            return res.defaultError(e);
         }
     }
 
     @Description("큐레이션 목록 상품 추가")
     @PostMapping("/{id}/add-product")
-    public ResponseEntity<CustomResponse> addProductToCuration(@PathVariable("id") Integer id,
-                                                               @RequestBody ArrayList<Integer> productIds) {
-        CustomResponse res = new CustomResponse();
+    public ResponseEntity<CustomResponse<List<CurationProductMap>>> addProductToCuration(@RequestHeader(value = "Authorization") Optional<String> auth,
+                                                                                         @PathVariable("id") Integer id,
+                                                                                         @RequestBody ArrayList<Integer> productIds) {
+        CustomResponse<List<CurationProductMap>> res = new CustomResponse();
+        Optional<TokenInfo> tokenInfo = jwt.validateAndGetTokenInfo(Set.of(TokenAuthType.ADMIN), auth);
+        if (tokenInfo == null) return res.throwError("인증이 필요합니다.", "FORBIDDEN");
         try {
             Optional<Curation> curation = curationService.findById(id);
             if (curation == null) throw new Error("큐레이션 정보를 찾을 수 없습니다.");
@@ -104,28 +105,28 @@ public class CurationController {
                 Product product = productService.selectProduct(productId);
                 if (product == null) throw new Error("상품 정보를 찾을 수 없습니다.");
             }
-            curationService.addProduct(id,productIds);
+            List<CurationProductMap> result = curationService.addProduct(id, productIds);
+            res.setData(Optional.ofNullable(result));
             return ResponseEntity.ok(res);
         } catch (Exception e) {
-            res.setIsSuccess(false);
-            res.setErrorMsg(e.getMessage());
-            return ResponseEntity.ok(res);
+            return res.defaultError(e);
         }
     }
 
     //DELETE
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<CustomResponse> deleteCuration(@PathVariable("id") Integer id) {
-        CustomResponse res = new CustomResponse();
+    public ResponseEntity<CustomResponse<Curation>> deleteCuration(@RequestHeader(value = "Authorization") Optional<String> auth,
+                                                                   @PathVariable("id") Integer id) {
+        CustomResponse<Curation> res = new CustomResponse();
+        Optional<TokenInfo> tokenInfo = jwt.validateAndGetTokenInfo(Set.of(TokenAuthType.ADMIN), auth);
+        if (tokenInfo == null) return res.throwError("인증이 필요합니다.", "FORBIDDEN");
         try {
             Optional<Curation> curation = curationService.findById(id);
             if (curation == null) throw new Error("큐레이션 데이터를 찾을 수 없습니다.");
             curationService.delete(id);
             return ResponseEntity.ok(res);
-        } catch (Error error) {
-            res.setIsSuccess(false);
-            res.setErrorMsg(error.getMessage());
-            return ResponseEntity.ok(res);
+        } catch (Exception e) {
+            return res.defaultError(e);
         }
     }
 }
