+++ src/main/java/com/matsinger/barofishserver/data/TipController.java
@@ -1,6 +1,9 @@
 package com.matsinger.barofishserver.data;
 
 import com.matsinger.barofishserver.banner.Banner;
+import com.matsinger.barofishserver.jwt.JwtService;
+import com.matsinger.barofishserver.jwt.TokenAuthType;
+import com.matsinger.barofishserver.jwt.TokenInfo;
 import com.matsinger.barofishserver.utils.Common;
 import com.matsinger.barofishserver.utils.CustomResponse;
 import com.matsinger.barofishserver.utils.S3.S3Uploader;
@@ -9,10 +12,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 @RestController
 @RequiredArgsConstructor
@@ -24,39 +24,40 @@ public class TipController {
     private final Common utils;
     private final S3Uploader s3;
 
+    private final JwtService jwt;
+
     @GetMapping("")
-    public ResponseEntity<CustomResponse> selectTipList() {
-        CustomResponse res = new CustomResponse();
+    public ResponseEntity<CustomResponse<List<Tip>>> selectTipList() {
+        CustomResponse<List<Tip>> res = new CustomResponse();
         try {
             List<Tip> tips = tipService.selectTipList();
             res.setData(Optional.ofNullable(tips));
             return ResponseEntity.ok(res);
         } catch (Exception e) {
-            res.setIsSuccess(false);
-            res.setErrorMsg(e.getMessage());
-            return ResponseEntity.ok(res);
+            return res.defaultError(e);
         }
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<CustomResponse> selectTip(@PathVariable("id") Integer id) {
-        CustomResponse res = new CustomResponse();
+    public ResponseEntity<CustomResponse<Tip>> selectTip(@PathVariable("id") Integer id) {
+        CustomResponse<Tip> res = new CustomResponse();
         try {
             Tip tip = tipService.selectTip(id);
             res.setData(Optional.ofNullable(tip));
             return ResponseEntity.ok(res);
         } catch (Exception e) {
-            res.setIsSuccess(false);
-            res.setErrorMsg(e.getMessage());
-            return ResponseEntity.ok(res);
+            return res.defaultError(e);
         }
     }
 
     @PostMapping("/add")
-    public ResponseEntity<CustomResponse> addTip(@RequestPart(value = "title") String title,
-                                                 @RequestPart(value = "description") String description,
-                                                 @RequestPart(value = "image") MultipartFile image) {
-        CustomResponse res = new CustomResponse();
+    public ResponseEntity<CustomResponse<Tip>> addTip(@RequestHeader(value = "Authorization") Optional<String> auth,
+                                                      @RequestPart(value = "title") String title,
+                                                      @RequestPart(value = "description") String description,
+                                                      @RequestPart(value = "image") MultipartFile image) {
+        CustomResponse<Tip> res = new CustomResponse();
+        Optional<TokenInfo> tokenInfo = jwt.validateAndGetTokenInfo(Set.of(TokenAuthType.ADMIN), auth);
+        if (tokenInfo == null) return res.throwError("인증이 필요합니다.", "FORBIDDEN");
         try {
             Tip tip = new Tip();
             title = utils.validateString(title, 100L, "제목");
@@ -69,18 +70,19 @@ public class TipController {
             res.setData(Optional.ofNullable(tipService.add(tip)));
             return ResponseEntity.ok(res);
         } catch (Exception e) {
-            res.setIsSuccess(false);
-            res.setErrorMsg(e.getMessage());
-            return ResponseEntity.ok(res);
+            return res.defaultError(e);
         }
     }
 
     @PostMapping("/update/{id}")
-    public ResponseEntity<CustomResponse> updateTip(@PathVariable("id") Integer id,
-                                                    @RequestPart(value = "title", required = false) String title,
-                                                    @RequestPart(value = "description", required = false) String description,
-                                                    @RequestPart(value = "image", required = false) MultipartFile image) {
-        CustomResponse res = new CustomResponse();
+    public ResponseEntity<CustomResponse<Tip>> updateTip(@RequestHeader(value = "Authorization") Optional<String> auth,
+                                                         @PathVariable("id") Integer id,
+                                                         @RequestPart(value = "title", required = false) String title,
+                                                         @RequestPart(value = "description", required = false) String description,
+                                                         @RequestPart(value = "image", required = false) MultipartFile image) {
+        CustomResponse<Tip> res = new CustomResponse();
+        Optional<TokenInfo> tokenInfo = jwt.validateAndGetTokenInfo(Set.of(TokenAuthType.ADMIN), auth);
+        if (tokenInfo == null) return res.throwError("인증이 필요합니다.", "FORBIDDEN");
         try {
             Tip tip = new Tip();
             tipService.selectTip(id);
@@ -99,22 +101,21 @@ public class TipController {
             res.setData(Optional.ofNullable(tipService.update(id, tip)));
             return ResponseEntity.ok(res);
         } catch (Exception e) {
-            res.setIsSuccess(false);
-            res.setErrorMsg(e.getMessage());
-            return ResponseEntity.ok(res);
+            return res.defaultError(e);
         }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<CustomResponse> deleteTip(@PathVariable("id") Integer id) {
+    public ResponseEntity<CustomResponse> deleteTip(@RequestHeader(value = "Authorization") Optional<String> auth,
+                                                    @PathVariable("id") Integer id) {
         CustomResponse res = new CustomResponse();
+        Optional<TokenInfo> tokenInfo = jwt.validateAndGetTokenInfo(Set.of(TokenAuthType.ADMIN), auth);
+        if (tokenInfo == null) return res.throwError("인증이 필요합니다.", "FORBIDDEN");
         try {
             tipService.delete(id);
             return ResponseEntity.ok(res);
         } catch (Exception e) {
-            res.setIsSuccess(false);
-            res.setErrorMsg(e.getMessage());
-            return ResponseEntity.ok(res);
+            return res.defaultError(e);
         }
     }
 }
